intelligence/requirements.txt
httpx
beautifulsoup4
langchain-text-splitters
aiofiles
chromadb
sentence-transformers
python-dotenv
fastapi
uvicorn
python-multipart
langchain-openai
pandas
pydantic
